Add input in cad links
