+++ addon/objects/uml-primitives/fd-uml-composition.js
@@ -4,6 +4,8 @@
 
 import FdUmlLink from './fd-uml-link';
 import { Link } from './fd-uml-link';
+import { MultiplicityView } from './links-view/fd-multiplicity-view';
+import joint from 'npm:jointjs';
 
 /**
   An object that describes a link of the composition type on the UML diagram.
@@ -19,7 +21,7 @@ export default FdUmlLink.extend({
     @method JointJS
   */
   JointJS() {
-    let properties = this.getProperties('id', 'repositoryObject', 'source', 'target', 'vertices', 'labels');
+    let properties = this.getProperties('id', 'repositoryObject', 'source', 'target', 'vertices', 'labels', 'startPoint', 'endPoint');
     return new Composition(properties);
   },
 });
@@ -34,5 +36,26 @@ export default FdUmlLink.extend({
   @constructor
 */
 export let Composition = Link.define('flexberry.uml.Composition', {
-  attrs: { '.marker-target': { d: 'M 26 10 L 13 17 L 0 10 L 13 3 z', fill: 'black' } },
+  attrs: {
+    '.marker-source': { d: 'M 26 10 L 13 17 L 0 10 L 13 3 z', fill: 'black' },
+    text: { visibility: 'hidden' },
+    rect: { visibility: 'hidden' }
+  }
+}, {
+  getLabelDistance: function (labelName, isVertical) {
+    switch (labelName) {
+      case 'startMultiplicity':
+      case 'startRole':
+        return 30;
+      case 'endMultiplicity':
+      case 'endRole':
+        return isVertical ? -10 : -5;
+      case 'description':
+        return 0.5;
+      default:
+        console.log('ERROR - choose correct label name');
+    }
+  }
 });
+
+joint.shapes.flexberry.uml.CompositionView = MultiplicityView;
