+++ addon/objects/uml-primitives/fd-uml-generalization.js
@@ -4,6 +4,8 @@
 
 import FdUmlLink from './fd-uml-link';
 import { Link } from './fd-uml-link';
+import { DescriptionView } from './links-view/fd-description-view';
+import joint from 'npm:jointjs';
 
 /**
   An object that describes a link of the inheritance type on the UML diagram.
@@ -18,7 +20,7 @@ export default FdUmlLink.extend({
     @method JointJS
   */
   JointJS() {
-    let properties = this.getProperties('id', 'repositoryObject', 'source', 'target', 'vertices', 'labels');
+    let properties = this.getProperties('id', 'repositoryObject', 'source', 'target', 'vertices', 'labels', 'startPoint', 'endPoint');
     return new Generalization(properties);
   },
 });
@@ -33,5 +35,10 @@ export default FdUmlLink.extend({
   @constructor
 */
 export let Generalization = Link.define('flexberry.uml.Generalization', {
-  attrs: { '.marker-target': { d: 'M 20 0 L 0 10 L 20 20 z', fill: 'white' } },
+  attrs: { '.marker-source': { d: 'M 20 0 L 0 10 L 20 20 z', fill: 'white' },
+  text: { visibility: 'hidden' },
+  rect: { visibility: 'hidden' }
+ }
 });
+
+joint.shapes.flexberry.uml.GeneralizationView = DescriptionView;
