+++ addon/objects/uml-primitives/fd-uml-nested-association.js
@@ -4,6 +4,8 @@
 
 import FdUmlLink from './fd-uml-link';
 import { Link } from './fd-uml-link';
+import { DescriptionView } from './links-view/fd-description-view';
+import joint from 'npm:jointjs';
 /**
   An object that describes a link of the association type on the UML diagram.
 
@@ -18,7 +20,7 @@ export default FdUmlLink.extend({
     @method JointJS
   */
   JointJS() {
-    let properties = this.getProperties('id', 'source', 'target', 'vertices', 'labels');
+    let properties = this.getProperties('id', 'source', 'target', 'vertices', 'labels', 'startPoint', 'endPoint');
     return new NestedClassAssociation(properties);
   },
 });
@@ -33,8 +35,13 @@ export default FdUmlLink.extend({
   @constructor
 */
 export let NestedClassAssociation = Link.define('flexberry.uml.NestedClassAssociation', {
-  attrs: { '.marker-target': { d: 'M 10, 10 ' +
+  attrs: { '.marker-source': { d: 'M 10, 10 ' +
                 'm -8, 0 ' +
                 'a 8,8 0 1,0 16,0 ' +
-                'a 8,8 0 1,0 -16,0 M 10 2 L 10 2 L 10 18', fill: 'transparent' } }
+                'a 8,8 0 1,0 -16,0 M 10 2 L 10 2 L 10 18', fill: 'transparent' },
+            text: { visibility: 'hidden' },
+            rect: { visibility: 'hidden' }
+  }
 });
+
+joint.shapes.flexberry.uml.NestedClassAssociationView = DescriptionView;
